+++ scs-client/C#/SCSSdkClient.Demo/SCSSdkClient.Demo/SCSSdkClientDemo.cs
@@ -12,6 +12,8 @@ namespace SCSSdkClient.Demo {
         /// </summary>
         public SCSSdkTelemetry Telemetry;
 
+        private float fuel;
+
         /// <inheritdoc />
         public SCSSdkClientDemo() {
             InitializeComponent();
@@ -67,11 +69,12 @@ namespace SCSSdkClient.Demo {
         private void TelemetryRefuelEnd(object sender, EventArgs e) =>  rtb_fuel.Invoke((MethodInvoker)(()=>rtb_fuel.BackColor = Color.Red));
 
         private void TelemetryRefuelPayed(object sender, EventArgs e) {
-            MessageBox.Show("Fuel Payed");
+            MessageBox.Show("Fuel Payed: " + fuel);
         }
 
 
-        private void Telemetry_Data(SCSTelemetry data, bool updated) { 
+        private void Telemetry_Data(SCSTelemetry data, bool updated) {
+            if (!updated) return;
             try {
                 if (InvokeRequired) {
                     Invoke(new TelemetryData(Telemetry_Data), data, updated);
@@ -130,7 +133,7 @@ namespace SCSSdkClient.Demo {
                 substances.Text = JsonConvert.SerializeObject(data.Substances, Formatting.Indented);
                 gameplayevent.Text = JsonConvert.SerializeObject(data.GamePlay, Formatting.Indented);
                 rtb_fuel.Text = data.TruckValues.CurrentValues.DashboardValues.FuelValue.Amount + " "+ data.SpecialEventsValues.Refuel ;
-               
+                fuel = data.GamePlay.RefuelEvent.Amount;
 
             } catch (Exception ex) {
                 // ignored atm i found no proper way to shut the telemetry down and down call this anymore when this or another thing is already disposed